src: Fix binding between mode and parameters in publisher
Previously, the code assumed that the correct parameters were provided
for each mode, but that’s not necessarily the case. If incorrect
parameters were provided (specifically, if the ones pertaining to the
given mode were missing), some NULL pointer dereferences would happen
and hence the code would crash.

Fix that by checking for the parameters and returning error 400 if they
are not present.

Coverity IDs: 1388546, 1388547

https://bugzilla.gnome.org/show_bug.cgi?id=776115